Refactor preset into RenderScript with support for hotkeys.
Preset is also a renderchain, so it can be used inside ScriptChain and
the like.
It's not yet possible to reuse Presets, this requires some care to avoid
loops.
