+++ Extensions/RenderScripts/Mpdn.Presets.cs
@@ -1,114 +1,118 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Windows.Forms;
-using Mpdn.Config;
-using Mpdn.PlayerExtensions.GitHub;
+using Mpdn.PlayerExtensions;
 using YAXLib;
 
 namespace Mpdn.RenderScript
 {
     namespace Mpdn.Presets
     {
-        public abstract class PresetRenderScriptBase : IRenderScriptUi
+        public class Selector<TOption> : RenderChain
+            where TOption : RenderChain, new()
         {
-            protected abstract RenderScriptPreset Preset { get; }
+            public List<TOption> Options { get; set; }
 
-            protected virtual IRenderScriptUi Script { get { return Preset.Script ?? RenderScript.Empty; } }
+            public int SelectedIndex { get; set; }
 
-            public virtual IRenderScript CreateRenderScript()
+            [YAXDontSerialize]
+            public TOption SelectedOption { get { return Options.ElementAtOrDefault(SelectedIndex); } }
+
+            public Selector()
             {
-                return Script.CreateRenderScript();
+                Options = new List<TOption>();
+                SelectedIndex = -1;
             }
 
-            public virtual void Destroy()
+            public override IFilter CreateFilter(IResizeableFilter sourceFilter)
             {
-                Script.Destroy();
-            }
-
-            public virtual void Initialize() { }
-
-            public virtual bool ShowConfigDialog(IWin32Window owner)
-            {
-                return Script.ShowConfigDialog(owner);
-            }
-
-            public virtual bool HasConfigDialog()
-            {
-                return Script.HasConfigDialog();
-            }
-
-            public virtual ExtensionUiDescriptor Descriptor
-            {
-                get 
-                {
-                    var descriptor = Script.Descriptor;
-                    descriptor.Guid = Preset.Guid;
-                    descriptor.Name = Preset.Name;
-                    return descriptor;
-                }
+                if (SelectedOption != null)
+                    return SelectedOption.CreateFilter(sourceFilter);
+                else
+                    return sourceFilter;
             }
         }
 
-        public class PresetRenderScript : PresetRenderScriptBase
+        public class RenderScriptPreset : RenderChain
         {
-            private RenderScriptPreset m_SavedPreset;
-            public RenderScriptPreset SavedPreset  
+            [YAXAttributeForClass]
+            public string Name { get; set; }
+
+            [YAXAttributeForClass]
+            public Guid Guid { get; set; }
+
+            public IRenderChainUi Script { get; set; }
+
+            public RenderScriptPreset()
             {
-                get { return m_SavedPreset; }
+                Guid = Guid.NewGuid();
+            }
+
+            public override IFilter CreateFilter(IResizeableFilter sourceFilter)
+            {
+                if (Script != null)
+                    return Script.GetChain().CreateFilter(sourceFilter);
+                else
+                    return sourceFilter;
+            }
+        }
+
+        public class PresetSettings : Selector<RenderScriptPreset> 
+        {
+            private Guid HotkeyGuid = Guid.NewGuid();
+            private string m_Hotkey;
+
+            public string Name { get; set; }
+            public string Hotkey 
+            {
+                get { return m_Hotkey; }
                 set
                 {
-                    m_SavedPreset = value;
-                    m_SavedPreset = PresetExtension.LoadPreset(m_SavedPreset.Guid) ?? m_SavedPreset;
-                }
+                    m_Hotkey = value;
+                    RegisterHotkey();
+                } 
             }
 
-            public PresetRenderScript(RenderScriptPreset preset)
+            public PresetSettings()
             {
-                SavedPreset = preset;
+                Name = "Preset Group";
+                Hotkey = "";
             }
 
-            protected override RenderScriptPreset Preset { get { return SavedPreset; } }
+            private void RegisterHotkey()
+            {
+                DynamicHotkeys.RegisterHotkey(HotkeyGuid, Hotkey, IncrementSelection);
+            }
+
+            private void IncrementSelection()
+            {
+                if (Options.Count > 0)
+                    SelectedIndex = (SelectedIndex + 1) % Options.Count;
+
+                if (SelectedOption != null)
+                    PlayerControl.ShowOsdText(Name + ": " + SelectedOption.Name);
+
+                PlayerControl.SetRenderScript(PlayerControl.ActiveRenderScriptGuid);
+            }
         }
 
-        public class PresetRenderChain : PresetRenderScript, IRenderChainUi
+        public class PresetRenderScript : RenderChainUi<PresetSettings, PresetDialog>
         {
-            public PresetRenderChain(RenderScriptPreset preset) : base(preset) 
+            protected override string ConfigFileName
             {
-                if (!(Script is IRenderChainUi)) throw new ArgumentException("Not a preset for a RenderChain");
-            }
-
-            public RenderChain GetChain()
-            {
-                return (Script as IRenderChainUi).GetChain();
-            }
-        }
-
-        public class ActivePresetRenderScript : PresetRenderScriptBase
-        {
-            private Guid m_Guid = new Guid("B1F3B882-3E8F-4A8C-B225-30C9ABD67DB1");
-
-            protected override RenderScriptPreset Preset 
-            {
-                get { return PresetExtension.ActivePreset ?? new RenderScriptPreset() { Script = RenderScript.Empty }; } 
-            }
-
-            public override void Initialize()
-            {
-                base.Initialize();
-
-                PresetExtension.ScriptGuid = m_Guid;
+                get { return "Mpdn.Presets"; }
             }
 
             public override ExtensionUiDescriptor Descriptor
             {
                 get
                 {
-                    var descriptor = base.Descriptor;
-                    descriptor.Name = "Preset";
-                    descriptor.Guid = m_Guid;
-                    descriptor.Description = "Active Preset: " + Preset.Name;
-                    return descriptor;
+                    return new ExtensionUiDescriptor
+                    {
+                        Name = Settings.Name ?? "Preset Group",
+                        Guid = new Guid("B1F3B882-3E8F-4A8C-B225-30C9ABD67DB1"),
+                        Description = (Settings.SelectedOption == null) ? "Group of Presets" : Settings.SelectedOption.Name
+                    };
                 }
             }
         }
